pyproject.toml
[tool.poetry]
name = "simple-news-rag"
version = "0.1.0"
description = "A minimal service to ingest, summarize, and search news articles (FastAPI, LangChain, PostgreSQL/pgvector)."
authors = ["Generated"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "~3.13"
# Web Framework & Runtime
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
pydantic-settings = "^2.3.4"

# Database (SQLAlchemy 2.x, AsyncPG, pgvector)
SQLAlchemy = {extras = ["asyncio"], version = "^2.0.43"}
alembic = "^1.16.5"
asyncpg = "^0.30.0"
pgvector = "^0.4.1"

# LangChain & OpenAI
langchain = "^0.3.27"
langchain-openai = "^0.3.32"
langchain-community = "^0.3.29"

# Loaders and Parsers
newspaper3k = "^0.2.8"
lxml = {extras = ["html_clean"], version = "^6.0.1"}
beautifulsoup4 = "^4.13.5" # Required by WebBaseLoader

# Utilities
numpy = "^2.0.0" # For embedding normalization
tenacity = "^8.5.0" # For robust retries
python-dateutil = "^2.9.0" # For robust date parsing
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.23.7"
httpx = "^0.27.0"
black = "^24.8.2"
isort = "^5.13.2"
mypy = "^1.13.0"
ruff = "^0.7.0"
types-python-dateutil = "^2.9.0.20250822"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py313']

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.13"
disallow_untyped_defs = false
ignore_missing_imports = true
warn_return_any = false
warn_unused_configs = true
show_error_codes = true